fix(#123): Fix BaseOption id type
- Changed the `id` type of `BaseOption` from `int` to `long`
  in `RawMaterialMapper`, `PreparationMapper`,
  `ProductIngredientMapper`, and
  `PreparationIngredientMapper`.

+++ src/main/java/org/samuel/storemanagement/domain/preparation/ingredient/mappers/PreparationIngredientMapper.java
@@ -47,7 +47,7 @@ public interface PreparationIngredientMapper {
         };
 
         return BaseOption.builder()
-                .id(unit.ordinal())
+                .id((long) unit.ordinal())
                 .name(name).build();
     }
