+++ src/main/java/org/samuel/storemanagement/domain/product/ingredient/mappers/ProductIngredientMapper.java
@@ -37,7 +37,7 @@ public interface ProductIngredientMapper {
         };
 
         return BaseOption.builder()
-                .id(unit.ordinal())
+                .id((long) unit.ordinal())
                 .name(name).build();
     }
